refactor: improve CSS cleaning and validation functions, enhance panel loading logic
